Refactorise le header plan d'action avec une modale pour changer le titre du plan
